feat: basic logger
